feat(sequencer/aggregation_consensus): Check signatures of votes
