+++ libs/feeds_processing/Cargo.toml
@@ -13,6 +13,7 @@ config = {path = "../config"}
 utils = {path = "../utils"}
 prometheus = {path = "../prometheus"}
 gnosis_safe = {path = "../gnosis_safe"}
+crypto = { path = "../../libs/crypto" }
 once_cell = "1.20.2"
 hex = "0.4.3"
 ringbuf = "0.4.7"
